Phase 5: Compose UI (NavGraph, 5 screens + VMs); fix Hilt 2.52→2.54 for Kotlin 2.1.0 metadata
Fix: alignment audio — continuous pan spectrum, suppress accuracy beacon during alignment, silence beyond 90°

+++ shared/src/commonMain/kotlin/com/boldexplorer/shared/audio/AudioCueEvent.kt
@@ -4,8 +4,8 @@ sealed class AudioCueEvent {
     // Continuous beacon: frequency maps to GPS accuracy (0m → 880 Hz, 30m → 220 Hz).
     data class AccuracyBeacon(val accuracyM: Double) : AudioCueEvent()
 
-    // Alignment ping at configurable Hz; pan = -1 (left), 0 (center), 1 (right).
-    data class AlignmentPing(val pan: Int, val aligned: Boolean) : AudioCueEvent()
+    // Alignment ping at configurable Hz; pan in [-1.0, 1.0] (-1 = full left, +1 = full right).
+    data class AlignmentPing(val pan: Float, val aligned: Boolean) : AudioCueEvent()
 
     // TTS announcement when a waypoint is reached.
     data class WaypointApproach(val waypointName: String) : AudioCueEvent()
